+++ app/src/main/java/com/zhangchong/toolsapplication/Data/SqlManager.java
@@ -4,12 +4,21 @@ import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
+import com.zhangchong.toolsapplication.Data.Bean.ExcelCellBean;
+import com.zhangchong.toolsapplication.Data.Bean.ExcelFileBean;
+import com.zhangchong.toolsapplication.Data.DAO.DaoEntrySchema;
+
 /**
  * Created by Zhangchong on 2015/4/4.
  */
 public class SqlManager extends SQLiteOpenHelper{
     private static final String DB_NAME = "tools.db";
     private static final int DB_VERSION = 1;
+
+    private static final DaoEntrySchema[] initTables = {
+            ExcelFileBean.schema,
+            ExcelCellBean.schema
+    };
 
     public SqlManager(Context context){
         super(context,DB_NAME, null, DB_VERSION);
@@ -18,12 +27,33 @@ public class SqlManager extends SQLiteOpenHelper{
     @Override
     public void onCreate(SQLiteDatabase db) {
         //创建表结构
+        onUpgrade(db, 0, DB_VERSION);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + DB_NAME);
-        onCreate(db);
+
+        if(oldVersion > newVersion){
+            db.execSQL("DROP TABLE IF EXISTS " + DB_NAME);
+            oldVersion = 0;
+        }
+
+        if(oldVersion < 1){
+            createBasicDB(db);
+            oldVersion = 1;
+        }
+
+        //TODO
+        if(oldVersion < 2){
+
+        }
+    }
+
+    private void createBasicDB(SQLiteDatabase db){
+        for (DaoEntrySchema schema : initTables) {
+            schema.dropTables(db);
+            schema.createTables(db);
+        }
     }
